pyproject.toml
[project]
name = "vaulter"
version = "0.1.0"
dependencies = [
  "cryptography>=43.0.0",
  "argon2-cffi>=23.1.0",
  "typer>=0.12.0",
  "structlog>=24.1.0",
  "pydantic>=2.7.0",
  "PyNaCl>=1.5.0"
]
[project.scripts]
vaulter = "vaulter.cli:app"

[tool.setuptools]
packages = ["vaulter"]
